requirements/requirements.txt
numpy
pandas
tensorflow
pydantic
strictyaml